feat: url filtering like clearURLs
Urls get automatically filtered through the clearURLs filters,
which almost produces the same result as with clearURLs.
There's also a config option to toggle it (enabled by default).
It looks like it laggs behind a little with the filtering,
but the filtering process itself takes very little time.
It just seems that it can't change the entry,
before the load has finished.
Yep it took that long, as i wasted my time with
jsmn, regex, pcre2, which none worked out..

Signed-off-by: chm46e <chm46e@duck.com>